Num mode tweaks
I decide that it might be nice to move shift onto a number
key to make it easier to press. With the left hand
4 isn't very pressable so switch to 5.

Also move the control keys to somewhere more
pressable in preference to the "symbol" keys.
0 can't really be touch-typed: but it seems that
minus can.

What is really going on here is that I'm adapting
keymappings to deal the fact that the keyboard is straight
but my hands want to come in at an angle (this is
exacerbated by the fact that I'm using a small
keyboard ). This makes some of the modifier keys
particularly difficult to type.  The effect is
likely still going to be present on most
keyboards, unless they are "ergonomic", and I don't
think I want to depend upon siting at a desk with a
massive keyboard.